Interfaz Roles y Permisos (sin editar roles)

+++ BL/AdministradordeSistema.cs
@@ -65,7 +65,7 @@ namespace BL
             }
         }
 
-        public void eliminarRol(int identity)
+        public void eliminarRol(long identity)
         {
             try
             {
